moves dockerfileextensions to separate class
